Fix adding UI to top of page
The originally checked `account-header` class seems to have been renamed. In the fallback, the UI is at the end instead. When I scroll down, I catch a glimpse of the Start/Stop area, but I have to keep scrolling down until the list is fully loaded before I can actually click on it. This fixes that and hopefully reduces the need to fix it in the future by always inserting the UI after the first div in the body.

Alternatively, the instructions could be updated.